requirements.txt
# ── Data collection ───────────────────────────────────────────────────────────
aiohttp==3.9.5
requests==2.32.3
shodan==1.31.0
python-nmap==0.7.1
dnspython==2.6.1
python-whois==0.9.4

# ── Async & utils ─────────────────────────────────────────────────────────────
asyncio==3.4.3
python-dotenv==1.0.1
pydantic==2.7.1
rich==13.7.1          # pretty terminal output
tqdm==4.66.4

# ── API server ────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn==0.30.1

# ── Reporting ─────────────────────────────────────────────────────────────────
jinja2==3.1.4
weasyprint==62.3       # PDF generation

# ── ML / fine-tuning (install separately, see README) ─────────────────────────
# torch==2.3.0+cu121
# transformers==4.41.2
# peft==0.11.1
# datasets==2.19.1
# trl==0.9.4
# unsloth                 # pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# bitsandbytes==0.43.1
